+++ tests/closures.rs
@@ -299,7 +299,7 @@ fn test_closures_shared_obj() -> Result<(), Box<EvalAltResult>> {
 
     // Make closure
     let f = move |p1: TestStruct, p2: TestStruct| -> Result<(), Box<EvalAltResult>> {
-        let action_ptr = res["action"].clone().cast::<FnPtr>();
+        let action_ptr = res["action"].clone_cast::<FnPtr>();
         let name = action_ptr.fn_name();
         engine.call_fn(&mut Scope::new(), &ast, name, (p1, p2))
     };